Fix bug in buf.c find_record
When matching a partial delimiter, the code checked whether the previous
buffer length was smaller than the remaining delimiter length.  This was
meant to check for unfindable delimiter spans but broke for delimiters
that were mostly in the second buffer.

This revision checks for sufficient bytes in the new buffer but will break
in the "unfindable" case listed.